* [jsfm] move element operation methods to operation.js
