+++ html5/runtime/vdom/operation.js
@@ -90,3 +90,82 @@ export function linkParent (node, parent) {
     linkParent(child, node)
   })
 }
+
+export function nextElement (node) {
+  while (node) {
+    if (node.nodeType === 1) {
+      return node
+    }
+    node = node.nextSibling
+  }
+}
+
+export function previousElement (node) {
+  while (node) {
+    if (node.nodeType === 1) {
+      return node
+    }
+    node = node.previousSibling
+  }
+}
+
+export function insertIndex (target, list, newIndex, changeSibling) {
+  if (newIndex < 0) {
+    newIndex = 0
+  }
+  const before = list[newIndex - 1]
+  const after = list[newIndex]
+  list.splice(newIndex, 0, target)
+  if (changeSibling) {
+    before && (before.nextSibling = target)
+    target.previousSibling = before
+    target.nextSibling = after
+    after && (after.previousSibling = target)
+  }
+  return newIndex
+}
+
+export function moveIndex (target, list, newIndex, changeSibling) {
+  const index = list.indexOf(target)
+  if (index < 0) {
+    return -1
+  }
+  if (changeSibling) {
+    const before = list[index - 1]
+    const after = list[index + 1]
+    before && (before.nextSibling = after)
+    after && (after.previousSibling = before)
+  }
+  list.splice(index, 1)
+  let newIndexAfter = newIndex
+  if (index <= newIndex) {
+    newIndexAfter = newIndex - 1
+  }
+  const beforeNew = list[newIndexAfter - 1]
+  const afterNew = list[newIndexAfter]
+  list.splice(newIndexAfter, 0, target)
+  if (changeSibling) {
+    beforeNew && (beforeNew.nextSibling = target)
+    target.previousSibling = beforeNew
+    target.nextSibling = afterNew
+    afterNew && (afterNew.previousSibling = target)
+  }
+  if (index === newIndexAfter) {
+    return -1
+  }
+  return newIndex
+}
+
+export function removeIndex (target, list, changeSibling) {
+  const index = list.indexOf(target)
+  if (index < 0) {
+    return
+  }
+  if (changeSibling) {
+    const before = list[index - 1]
+    const after = list[index + 1]
+    before && (before.nextSibling = after)
+    after && (after.previousSibling = before)
+  }
+  list.splice(index, 1)
+}
